+++ src/Fields/Types/AbstractField.php
@@ -245,6 +245,10 @@ abstract class AbstractField
             }
 
             if ($locale && $this->isLocalized()) {
+                if (method_exists($model, 'isDynamicKey') && $model->isDynamicKey($this->getColumn())) {
+                    return $model->dynamic($this->getColumn(), $locale);
+                }
+
                 return $model->getTranslationFor($this->getColumn(), $locale);
             }